Add an investment boost if a player increases investment in a point cluster
A point cluster that's <24 hours old is less likely to be useful than
one that's been around for two or more days of work.

i.e. This is likely to be the difference between a temporary "mining trip base"
and a primary base, respectively. We should provide stronger protection for
the primary base.